Add dummy file to satisfy requirements
common_path.sh is a script that normally sets PATH based on KALDI_ROOT
to find kaldi executables. We don't need it, because in our case kaldi
is installed in /opt and all binaries sit in /opt/kaldi/bin. However,
some of the scripts Rong uses still make use of the common_path, so we
just fake it.